# py-genai/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hello-GenAI API",
    "description": "REST API for the Hello-GenAI local chat application powered by Docker Model Runner.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "components": {
    "parameters": {
      "session_id": {
        "name": "session_id",
        "in": "path",
        "required": true,
        "description": "UUID of the session",
        "schema": { "type": "string", "format": "uuid" },
        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      },
...
    },
    "schemas": {
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"], "summary": "List projects",
        "responses": { "200": { "description": "Projects with session counts" } }
      },
      "post": {
        "tags": ["Projects"], "summary": "Create a project",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["name"],
          "properties": { "name": { "type": "string" }, "system_prompt": { "type": "string", "nullable": true } }
        }, "example": { "name": "Work", "system_prompt": null } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Missing name" } }
      }
    },
    "/api/projects/{project_id}": {
      "parameters": [ { "name": "project_id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 } ],
      "patch": {
        "tags": ["Projects"], "summary": "Rename or set a project's system prompt",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "properties": { "name": { "type": "string" }, "system_prompt": { "type": "string", "nullable": true } }
        } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Projects"], "summary": "Delete a project",
        "description": "Its sessions become unfiled (project_id → null); scoped memories and documents are removed.",
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
    "/api/sessions/{session_id}/project": {
      "post": {
        "tags": ["Projects"], "summary": "Assign a session to a project",
        "parameters": [ { "$ref": "#/components/parameters/session_id" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "properties": { "project_id": { "type": "integer", "nullable": true } }
        }, "example": { "project_id": 1 } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
    "/api/documents": {
      "get": {
        "tags": ["Documents"], "summary": "List knowledge-base documents",
        "parameters": [ { "name": "project_id", "in": "query", "required": false, "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "Documents + embeddings_available flag" } }
      },
      "post": {
        "tags": ["Documents"], "summary": "Ingest a document",
        "description": "Multipart PDF/text file upload (`file`) or JSON `{ filename, text }`. Optional `project_id`. Chunked and embedded when embeddings are configured.",
        "requestBody": { "required": true, "content": {
          "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" }, "project_id": { "type": "integer" } } } },
          "application/json": { "schema": { "type": "object", "properties": { "filename": { "type": "string" }, "text": { "type": "string" }, "project_id": { "type": "integer" } } } }
        } },
        "responses": { "201": { "description": "Ingested (document_id, chunks, embedded)" }, "400": { "description": "No text found" } }
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": ["Documents"], "summary": "Remove a document",
        "parameters": [ { "name": "document_id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
...
  }
}